pkg/helm/engine: Handle empty templates
The code adding `ownerReference` information to objects generated by
Helm templates would incorrectly handle 'empty' template renderings,
which are valid (and common) in charts, e.g. when certain objects are
only to be deployed based on some `values` (e.g. RBAC, `Ingress`es,...).

When such rendering occurred (an empty string, or maybe a string with
only some comment lines), the code would parse this as YAML, which would
yield an empty dictionary, then add a `metadata` field with the
`ownerReference` data in it, then pass this on to be deployed.

The above is of course flawed: we'd all of a sudden try to PUT an object
which has no `Kind`, no `apiVersion`,... only a `metadata` field, which
would then be rejected by the Kubernetes API server.

To fix this, empty template renderings are now skipped.